download handled and data type performed
