+++ src/utils/__init__.py
@@ -1,1 +1,2 @@
-from .pipeline import Pipeline
+from src.utils.pipeline import Pipeline
+from src.utils.files import auxiliar_files
